Enforce settings for newly created -> saved files
When users create a new file, it doesn't have a name yet so the plugin
doesn't know which settings it should apply. Later, when users save the
new file and assign a name to it, they expect the plugin to apply the
settings from their `.editorconfig` file according to the new file name
and it doesn't happen so they have to close the file and open it again.
Very annoying. This PR fixes the problem by applying the settings via
`on_post_save` hook. To make sure we don't do that for every save
operation (as it might lead to performance degradation), we mark the
files and skip them later.