Also run on_load
on_activated is only called when a view gains focus. When using Ctrl+P
to look for a file the indent settings are not applied until a file is
selected. This results in the settings changing when a file is finally
picked. By applying on the load event, the settings are applied earlier
and you don't visually notice them being applied.